refactor: restructure readme API handling and modularize constants
