+++ src/inngest/constants.ts
@@ -1,0 +1,3 @@
+export const INNGEST_EVENT = 'app/readme-develop/process';
+
+export const README_API_BASE_URL = 'https://dash.readme.com/api/v1';
